+++ frontend/src/components/Hero.tsx
@@ -3,6 +3,7 @@ import {
   useState,
   useRef,
 } from 'react';
+import Head from 'next/head';
 import { CldImage } from 'next-cloudinary';
 
 import styles from './hero.module.scss';
@@ -10,16 +11,20 @@ import styles from './hero.module.scss';
 export default function Hero({
   imagePath,
   alt,
+  heroImageData,
 }: {
   imagePath: string,
   alt: string,
+  heroImageData: any,
 }) {
   const [heroDimensions, setHeroDimensions] = useState({
     height: undefined,
     width: undefined,
   });
-  
+
   const ref: any = useRef(undefined);
+
+  const currentHero = heroImageData && heroImageData.filter((hero: any) => hero.public_id === imagePath)[0];
 
   useEffect(() => {
     const { current }: any = ref;
@@ -35,6 +40,11 @@ export default function Hero({
       ref={ref}
       className={styles.hero}
     >
+      {currentHero && (
+        <Head>
+          <link rel="preload" fetchpriority="high" as="image" href={currentHero.secure_url} type={`${currentHero.resource_type}/${currentHero.format}`} />
+        </Head>
+      )}
       {heroDimensions.height && heroDimensions.width && (
         <CldImage
           width={heroDimensions.width}