+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [7.1.9] - 2024-02-03
+### Updated
+- Fix accessibility issue on google map
+- Adjust hero to preload image
+
 ## [7.1.8] - 2024-02-03
 ### Updated
 - Fix css on google map for mobile